feat(codex): split bugs into input/algo/out
tried to make a scuffed way to check the bug accuracy
on different types of sections of code. for example
input/output/algorithm. this is based on the line that
was changed. so if the line contains input functions then
we mark it as input, the same for output but with print.
otherwise is considered algorithm part